Flush embedded channel.
Motivation:

apple/swift-nio#421 changed the behaviour of EmbeddedChannel to no longer
automatically succeed write promises without flushes. This is a reasonable
behavioural change, but broke a test that was working around but still
implicitly relied on the old behaviour.

Modifications:

Add the missing flushes.

Result:

This test will pass both on NIO master and NIO 1.7.0