Correctly keep buffers alive.
Motivation:

BIO_new_mem_buf does not take a copy of its underlying memory, so it's
enormously important that the pointer we pass it remain valid for the
duration of the lifetime of the BIO. This was previously not actually
something we did. Very bad.

Modifications:

Extended the scope of the `withUnsafeBytes` call until the `BIO` created
from those bytes is supposed to go out of scope.

Result:

Better memory safety.