Support extracting public keys in handshakes. (#49)
Motivation:

Users have requested the ability to extract information about the
public keys provided by peers during TLS handshakes. This is a reasonable
request: it's potentially very useful to have access to this information.

While I'm reluctant to add too much in the way of general cryptographic
functionality, it seems like a good idea to enable it for others. For
this reason, this patch extends NIOOpenSSL to enable this use case.

Modifications:

1. Added an OpenSSLPublicKey class that can be serialized to bytes.
2. Added support for obtaining an OpenSSLPublicKey from an
    OpenSSLCertificate.
3. Added support for setting the OpenSSL cert verification callback.

Result:

Much more flexibility and power in the library.