Support removing TLS handlers from live connections. (#54)
Motivation:

In some unusual protocol scenarios it is desirable to remove a TLS
tunnel from an active connection. Currently NIOOpenSSL does not
support doing this: the only way to get it to issue a CLOSE_NOTIFY
is to call close() on the Channel. This is less than ideal.

Modifications:

- Added a OpenSSLHandler.stopTLS method
- Reworked the internal state machine to handle this use-case.

Result:

Users can safely unwrap TLS from a Channel.