Clean up now-unnecessary warnings.
Motivation:

As we bumped the dependency of NIO to 1.12, we can now avoid some
deprecation warnings, as even the low-end of NIO versions we
support will be safe.

While we're here, we can remove the warnings that Swift 5 started
emitting for our visibility specifiers in extensions.

Modifications:

- Replaced changeCapacity with reserveCapacity
- Replaced numThreads with numberOfThreads
- Removed redundant visibility specifiers

Result:

Back to warning free compilation.