Namespace header-only inline functions. (#79)
Motivation:

Our previous namespacing effort used symbol analysis of a compiled binary
to locate all the symbols that needed to be namespaced away. This works
fine for regular C and C++ symbols, but fails for any header-only
inline functions.

These functions, while not emitted into the binary, can cause problems
with the C++ ODR if they are included in multiple clang modules. This
unfortunately does happen if the user includes both this library and,
for example, modularised headers that include OpenSSL headers.

Fortunately, we can namespace the inline functions as well.

Modifications:

- Namespaced all the inline headers by source analysis.

Result:

Fewer code conflicts.