Record BoringSSL revision hash
Motivation:

When debugging, users may find it helpful to be able to nail down exactly
which version of BoringSSL is used to generate the specific release. We should
record this in two easy-to-find places: one in the Package.swift for users
going to GitHub, and one in the Sources tree for users who check the code
out.

Modifications:

- Updated the script to record the BoringSSL hash.

Result:

Easier to find out what BoringSSL release is in use.