Remove the static target from regular Package.swift. (#85)
Motivation:

We have a static library target that is used as part of the BoringSSL
vendoring setup to generate a list of hidden symbols. This works nicely,
but building it fires out a bunch of warnings due to files with no symbols
in them. That's fine during the build phase, but most developers should
never need to see it.

Modifications:

- Hide the static archive product most of the time.
- Update vendor script to un-hide the static archive product on vendoring.
- Enable warnings-as-errors.

Result:

Fewer warnings.