Expose callback for SSLKEYLOGFILE.
Motivation:

In many cases it is useful to support dumping the keys used in a TLS
connection, primarily for debugging purposes. BoringSSL provides a
callback that can be used for this purpose: we should translate it into
Swift and expose it to our users.

Modifications:

- Added a keyLogCallback parameter to TLSConfiguration.
- Wired that callback up to BoringSSL.

Result:

Building blocks of SSLKEYLOGFILE support are present.