Unwrapping timeout shouldn't close channel.
Motivation:

If the unwrapping timeout fires, rather than close the channel we should
just fire an error down the pipeline. This is in-line with standard NIO
behaviour, and allows users to opt-out of channel closure in the event
that they need it.

Modifications:

- Internal state modification to have different behaviour when
    unwrapping and closing timeout fires.

Result:

Users can control what happens if unwrapping times out.