Stop using commonName to validate certificates.
Motivation:

CommonName as part of certificate hostname verification has been
deprecated for more than two decades. Browsers have finally stopped
using the commonName as part of verification, and the WebPKI forbids
putting names into the commonName that are not also part of the
subjectAlternativeName extension.

Continuing to trust the commonName as part of the certificate hostname
will put us in a weaker position than most of the browser ecosystem.
That's not a good place to be, so we should flip our default to not
trusting the commonName any longer.

For pragmatic reasons, we do allow an override, but the default
configuration is off. Users using this override will likely find
themselves disappointed when we remove it at some point in the future.

Modifications:

- Added TLSConfiguration options to enable trusting commonName fields
  for hostname verification.
- Made investigating the commonName field optional during hostname
  verification.
- Tests to validate the new behaviour.

Result:

No trusting common names by default anymore.