issue-231: Adding CA Names to the TLS handshake.
Motivation:

This is an enhancement to add support for send CA names during the handshake for #231.

Modifications:

Changes to:
TLSConfiguration.swift: Added the `sendCANameList` flag to indicate loading the CA names on `.fullVerification`.
SSLContext.swift: Added the CA Names derived from the `trustRoots`.
TLSConfigurationTests.swift: Added the test case for `testFullVerificationWithCANames`.
NIOSSLTestHelpers.swift: Added a new set of PEMs for TLS and client authentication testing.

Result:

Addition for #231.