Fix race caused by setting .inputClosed
Motivation:

Previously we did not check the current state when setting the
`NIOSSLHandler`'s state to `.inputClosed`. This resulted in an issue
where when we were in state `.closing`, then set the state to
`.inputClosed`, we would end up with a fatalError in our shutdown /
scheduledShutdown as state `.inputClosed` is not a shutting-down state.

Modifications:

* check current state before setting `state = .inputClosed`,
  specifically not setting the state to `.inputClosed` when we are
  already in the process of doing a full closure