Pass empty psk hint to client callback when server doesn't provide one
Motivation:

The actual implementation of the clientPSKCallback and serverPSKCallback
makes the server psk hint mandatory.

Reading RFC4279 it seems that, on the opposite, the server psk hint
should be provided only on application that mandate it.

https://datatracker.ietf.org/doc/html/rfc4279#section-5.2

Modifications:

I modified clientPSKCallback method to fallback to an empty string
when no psk hint is provided by the server instead of returning without
providing a psk identity.
Following the same logic, the serverPSKCallback fallbacks to an empty
string when none is provided by the server configuration.

Result:

After my change, the NIOSSLContext.pskClientConfigurationCallback will
be called with an empty string when the server doesn't provide any hint.

The NIOSSLContext.pskServerConfigurationCallback will
be called with an empty string when the config doesn't provide any hint.

This allow client which doesn't expect any server hint to provide a
PSK identity.
This allow server which doesn't want to provide an hint to do so.