Add documentation of how trust roots settings affect validators
Motivation:

The trust roots settings have sometimes got non-obvious impacts
on which validator is used. These are a legacy behaviour and so
changing them is likely to cause services to encounter
challenging bugs, so it's probably not worth fixing this behaviour
without doing a deprecate-and-replace cycle.

A deprecate-and-replace of TLSConfiguration is going to be
necessary in the near-term because it is a suboptimal structure that
is carrying a lot of load. For now, however, let's just make it
possible for users to diagnose what is going on.

Modifications:

- Document the existing behaviour

Result:

Users have a chance of figuring out what is going on.