Enable opportunistic TLS verification (#565)
### Motivation:

In some scenarios a server might want to accept connections from both,
authenticated and not authenticated peers and defer decisions after the
handshake. This new options allows a host to only do TLS verificaiton if
the peer presents certificates and accept connections otherwise.

### Modifications:

Add a new associated value to the `none` case of
`CertificationVerification` to maintain backwards compatibility. This
option enables opportunistic verification.

### Result:

A new configuration case for TLS verification.

---------

Co-authored-by: aryan-25 <aryan_shah@apple.com>